+++ src/tag/parser/compile/object.cpp
@@ -10,7 +10,7 @@ namespace Invader::Parser {
         tag.render_bounding_radius = tag.render_bounding_radius < tag.bounding_radius ? tag.bounding_radius : tag.render_bounding_radius;
     }
 
-    template <typename T> void compile_object(T &tag) {
+    template <typename T> void compile_object(T &tag, BuildWorkload &workload, std::size_t tag_index) {
         tag.scales_change_colors = 0;
         for(auto &c : tag.change_colors) {
             if(c.scale_by != HEK::FunctionScaleBy::FUNCTION_SCALE_BY_NONE) {
@@ -19,6 +19,11 @@ namespace Invader::Parser {
             }
         }
         fix_render_bounding_radius(tag);
+
+        // Animate what?
+        if(tag.model.path.size() == 0 && tag.animation_graph.path.size() != 0) {
+            workload.report_error(BuildWorkload::ErrorType::ERROR_TYPE_ERROR, "Object tag has a model tag but no animation graph", tag_index);
+        }
     }
 
     void ObjectChangeColors::postprocess_hek_data() {
@@ -146,9 +151,9 @@ namespace Invader::Parser {
         tag.hard_ping_interrupt_ticks = static_cast<std::int16_t>(tag.hard_ping_interrupt_time * TICK_RATE);
     }
 
-    void Biped::pre_compile(BuildWorkload &, std::size_t, std::size_t, std::size_t) {
+    void Biped::pre_compile(BuildWorkload &workload, std::size_t tag_index, std::size_t, std::size_t) {
         this->object_type = HEK::ObjectType::OBJECT_TYPE_BIPED;
-        compile_object(*this);
+        compile_object(*this, workload, tag_index);
         compile_unit(*this);
         this->cosine_stationary_turning_threshold = std::cos(this->stationary_turning_threshold);
         this->crouch_camera_velocity = 1.0f / this->crouch_transition_time / TICK_RATE;
@@ -158,14 +163,14 @@ namespace Invader::Parser {
         this->sine_uphill_falloff_angle = static_cast<float>(std::sin(this->uphill_falloff_angle));
         this->sine_uphill_cutoff_angle = static_cast<float>(std::sin(this->uphill_cutoff_angle));
     }
-    void Vehicle::pre_compile(BuildWorkload &, std::size_t, std::size_t, std::size_t) {
+    void Vehicle::pre_compile(BuildWorkload &workload, std::size_t tag_index, std::size_t, std::size_t) {
         this->object_type = HEK::ObjectType::OBJECT_TYPE_VEHICLE;
-        compile_object(*this);
+        compile_object(*this, workload, tag_index);
         compile_unit(*this);
     }
     void Weapon::pre_compile(BuildWorkload &workload, std::size_t tag_index, std::size_t, std::size_t) {
         this->object_type = HEK::ObjectType::OBJECT_TYPE_WEAPON;
-        compile_object(*this);
+        compile_object(*this, workload, tag_index);
 
         // Jason jones autoaim for the rocket warthog
         if(workload.building_stock_map && (workload.tags[tag_index].path == "vehicles\\rwarthog\\rwarthog_gun")) {
@@ -181,31 +186,31 @@ namespace Invader::Parser {
             }
         }
     }
-    void Equipment::pre_compile(BuildWorkload &, std::size_t, std::size_t, std::size_t) {
+    void Equipment::pre_compile(BuildWorkload &workload, std::size_t tag_index, std::size_t, std::size_t) {
         this->object_type = HEK::ObjectType::OBJECT_TYPE_EQUIPMENT;
-        compile_object(*this);
+        compile_object(*this, workload, tag_index);
     }
-    void Garbage::pre_compile(BuildWorkload &, std::size_t, std::size_t, std::size_t) {
+    void Garbage::pre_compile(BuildWorkload &workload, std::size_t tag_index, std::size_t, std::size_t) {
         this->object_type = HEK::ObjectType::OBJECT_TYPE_GARBAGE;
-        compile_object(*this);
+        compile_object(*this, workload, tag_index);
     }
-    void Projectile::pre_compile(BuildWorkload &, std::size_t, std::size_t, std::size_t) {
+    void Projectile::pre_compile(BuildWorkload &workload, std::size_t tag_index, std::size_t, std::size_t) {
         this->object_type = HEK::ObjectType::OBJECT_TYPE_PROJECTILE;
-        compile_object(*this);
+        compile_object(*this, workload, tag_index);
         this->initial_velocity /= TICK_RATE;
         this->final_velocity /= TICK_RATE;
     }
-    void Scenery::pre_compile(BuildWorkload &, std::size_t, std::size_t, std::size_t) {
+    void Scenery::pre_compile(BuildWorkload &workload, std::size_t tag_index, std::size_t, std::size_t) {
         this->object_type = HEK::ObjectType::OBJECT_TYPE_SCENERY;
-        compile_object(*this);
+        compile_object(*this, workload, tag_index);
     }
-    void Placeholder::pre_compile(BuildWorkload &, std::size_t, std::size_t, std::size_t) {
+    void Placeholder::pre_compile(BuildWorkload &workload, std::size_t tag_index, std::size_t, std::size_t) {
         this->object_type = HEK::ObjectType::OBJECT_TYPE_PLACEHOLDER;
-        compile_object(*this);
+        compile_object(*this, workload, tag_index);
     }
-    void SoundScenery::pre_compile(BuildWorkload &, std::size_t, std::size_t, std::size_t) {
+    void SoundScenery::pre_compile(BuildWorkload &workload, std::size_t tag_index, std::size_t, std::size_t) {
         this->object_type = HEK::ObjectType::OBJECT_TYPE_SOUND_SCENERY;
-        compile_object(*this);
+        compile_object(*this, workload, tag_index);
     }
 
     template <typename T> void compile_device(T &tag) {
@@ -215,20 +220,20 @@ namespace Invader::Parser {
         tag.inverse_position_acceleration_time = 1.0f / (TICK_RATE * tag.position_acceleration_time);
     }
 
-    void DeviceMachine::pre_compile(BuildWorkload &, std::size_t, std::size_t, std::size_t) {
+    void DeviceMachine::pre_compile(BuildWorkload &workload, std::size_t tag_index, std::size_t, std::size_t) {
         this->object_type = HEK::ObjectType::OBJECT_TYPE_DEVICE_MACHINE;
-        compile_object(*this);
+        compile_object(*this, workload, tag_index);
         compile_device(*this);
         this->door_open_time_ticks = static_cast<std::uint32_t>(this->door_open_time * TICK_RATE);
     }
-    void DeviceControl::pre_compile(BuildWorkload &, std::size_t, std::size_t, std::size_t) {
+    void DeviceControl::pre_compile(BuildWorkload &workload, std::size_t tag_index, std::size_t, std::size_t) {
         this->object_type = HEK::ObjectType::OBJECT_TYPE_DEVICE_CONTROL;
-        compile_object(*this);
+        compile_object(*this, workload, tag_index);
         compile_device(*this);
     }
-    void DeviceLightFixture::pre_compile(BuildWorkload &, std::size_t, std::size_t, std::size_t) {
+    void DeviceLightFixture::pre_compile(BuildWorkload &workload, std::size_t tag_index, std::size_t, std::size_t) {
         this->object_type = HEK::ObjectType::OBJECT_TYPE_DEVICE_LIGHT_FIXTURE;
-        compile_object(*this);
+        compile_object(*this, workload, tag_index);
         compile_device(*this);
     }
 
@@ -363,9 +368,6 @@ namespace Invader::Parser {
         if(struct_val.animation_graph.tag_id.read().is_null()) {
             workload.report_error(BuildWorkload::ErrorType::ERROR_TYPE_WARNING, "Biped has no animation graph, so the biped will not spawn", tag_index);
         }
-        else if(model_id.is_null()) {
-            workload.report_error(BuildWorkload::ErrorType::ERROR_TYPE_ERROR, "Biped is missing a model", tag_index);
-        }
         else {
             auto &model_tag = workload.tags[model_id.index];
             auto &model_tag_header = workload.structs[*model_tag.base_struct];
@@ -380,9 +382,6 @@ namespace Invader::Parser {
                     }
                 }
             }
-            else {
-                workload.report_error(BuildWorkload::ErrorType::ERROR_TYPE_ERROR, "Biped model has no nodes", tag_index);
-            }
         }
 
         head_index = this->head_model_node_index;
@@ -393,9 +392,6 @@ namespace Invader::Parser {
 
         if(struct_val.animation_graph.tag_id.read().is_null()) {
             workload.report_error(BuildWorkload::ErrorType::ERROR_TYPE_WARNING, "Vehicle has no animation graph, so the vehicle will not spawn", tag_index);
-        }
-        else if(struct_val.model.tag_id.read().is_null()) {
-            workload.report_error(BuildWorkload::ErrorType::ERROR_TYPE_ERROR, "Vehicle is missing a model", tag_index);
         }
 
         calculate_object_predicted_resources(workload, struct_index);